Changed recRate in Recovered from int to the correct type float. Added rate of indivuals retured from recovered pop to susceptible pop.

+++ Recovered.cs
@@ -3,7 +3,8 @@
 public class Recovered
 {
 	int recPop;
-	int recRate;
+	float recRate;
+
 
 	public Recovered()
 	{
